commit by bash

+++ fs/ide.c
@@ -78,11 +78,11 @@ int time_read() {
 	} while (time == 0);
 */	
 	int a = 999;	
-	syscall_read_dev((u_int) &time, read, 4);
+	syscall_read_dev(&time, read, 4);
 	while (time == 0) {
 		//syscall_read_dev((u_int) &time, trigger, 4);
-		syscall_write_dev((u_int) &a, trigger, 4);
-		syscall_read_dev((u_int) &time, read, 4);
+		syscall_write_dev(&a, trigger, 4);
+		syscall_read_dev(&time, read, 4);
 	} 
 	return time;
 }
